pyproject.toml
[project]
name = "fakeShop"
version = "0.1.0"
description = "None"
readme = "README.md"
authors = [{ name = "pro2015", email = "" }]
requires-python = ">=3.12"
dependencies = [
    "fletxr==0.1.4.a2",
    "flet[all]>=0.28.3",
    "httpx>=0.28.1",
    "fletcarousel>=0.0.14",
]
